Move ParserTestCase and its instances into dedicated classes
